Add Content model and refactor SecureStream
Added a new Content model to represent different types of content. Refactored SecureStream to use `@Model` and adjusted related code for profile relationship and content handling. Removed SecureStreamModel and related queries for cleaner implementation.
