+++ Arkavo/Arkavo/SecureStream.swift
@@ -2,90 +2,52 @@ import AppIntents
 import Foundation
 import SwiftData
 
-struct SecureStream: AppEntity, Identifiable, Codable {
-    static var typeDisplayRepresentation: TypeDisplayRepresentation = "SecureStream"
-    static var defaultQuery = SecureStreamQuery()
-
-    var id: UUID
+@Model
+final class SecureStream: Identifiable {
+    @Attribute(.unique) var id: UUID
     var name: String
-    var streamDescription: String
     var createdAt: Date
     var updatedAt: Date
     var tags: [String]
     var ownerID: UUID
-    var contents: [Content]
+    @Relationship var profile: Profile?
+
+    init(id: UUID = UUID(), name: String, ownerID: UUID) {
+        self.id = id
+        self.name = name
+        self.createdAt = Date()
+        self.updatedAt = Date()
+        self.tags = []
+        self.ownerID = ownerID
+    }
+}
+
+extension SecureStream: AppEntity {
+    static var typeDisplayRepresentation: TypeDisplayRepresentation = "SecureStream"
+    static var defaultQuery = SecureStreamQuery()
 
     var displayRepresentation: DisplayRepresentation {
         DisplayRepresentation(title: "\(name)")
     }
 
-    init(id: UUID = UUID(), name: String, streamDescription: String, ownerID: UUID, isPublic _: Bool = false) {
-        self.id = id
-        self.name = name
-        self.streamDescription = streamDescription
-        createdAt = Date()
-        updatedAt = Date()
-        tags = []
-        self.ownerID = ownerID
-        contents = []
-    }
-
     static var typeDisplayName: String {
         "SecureStream"
     }
+}
 
-    // TODO: revisit metadata
-    mutating func shareContent(type: ContentType, data: ContentData, metadata: [String: String], createdBy userID: UUID) -> Result<Content, ContentError> {
-        let newContent = Content(id: UUID(), type: type, data: data, metadata: metadata, createdAt: Date(), createdBy: userID)
-        contents.append(newContent)
-        updatedAt = Date()
+struct SecureStreamQuery: EntityQuery {
+    typealias Entity = SecureStream
 
-        return .success(newContent)
+    func entities(for identifiers: [SecureStream.ID]) async throws -> [SecureStream] {
+        // Implement this method to fetch SecureStream instances
+        // This is just a placeholder implementation
+        []
     }
-}
 
-enum ContentError: Error {
-    case notAMember
-    case contentNotFound
-    case notAuthorized
-}
-
-// Define various content types
-enum ContentType: String, Codable {
-    case text
-    case image
-    case video
-    case audio
-    case document
-    case link
-}
-
-struct Content: Identifiable, Codable {
-    let id: UUID
-    let type: ContentType
-    let data: ContentData
-    let metadata: [String: String]
-    let createdAt: Date
-    let createdBy: UUID
-}
-
-struct ContentData: Codable {
-    let title: String
-    let description: String?
-    let url: URL?
-    let textContent: String?
-    let fileSize: Int64?
-    let duration: TimeInterval?
-}
-
-@Model
-final class SecureStreamModel: ObservableObject {
-    @Attribute(.unique) var id: UUID
-    var stream: SecureStream
-
-    init(stream: SecureStream) {
-        id = stream.id
-        self.stream = stream
+    func suggestedEntities() async throws -> [SecureStream] {
+        // Implement this method to suggest SecureStream instances
+        // This is just a placeholder implementation
+        []
     }
 }
 
@@ -109,21 +71,5 @@ struct SecureStreamAppIntent: AppIntent {
 
     static var parameterSummary: some ParameterSummary {
         Summary("View the secure stream with ID \(\.$streamIDString)")
-    }
-}
-
-struct SecureStreamQuery: EntityQuery {
-    typealias Entity = SecureStream
-
-    func entities(for _: [SecureStream.ID]) async throws -> [SecureStream] {
-        // Implement this method to fetch SecureStream instances
-        // This is just a placeholder implementation
-        []
-    }
-
-    func suggestedEntities() async throws -> [SecureStream] {
-        // Implement this method to suggest SecureStream instances
-        // This is just a placeholder implementation
-        []
     }
 }
